add init window
now we have a link between the init window and the main window through login/logout. login currently has no username/password logic but that can be changed as soon as we have a db connection. quit always exists the application no matter where it is clicked (init or main). the plan is to have another window named Results Section which will show the result of the given query into the database, and will have options {go back to main window, logout, quit}.